# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HealthDesk Plus API",
        "description": "Comprehensive healthcare management system for doctors, patients, appointments, prescriptions, and medical records.",
        "contact": {
            "name": "HealthDesk Plus Support",
            "email": "support@healthdeskplus.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.healthdeskplus.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/ai/translate": {
            "post": {
                "tags": [
                    "Language Support"
                ],
                "summary": "Translate text to target language",
                "description": "Translate text using Google Translate API",
                "operationId": "ed489a0d36f140968214a3571f2c1688",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text",
                                    "target_language"
                                ],
                                "properties": {
                                    "text": {
                                        "description": "Text to translate",
                                        "type": "string",
                                        "example": "Hello, how are you?"
                                    },
                                    "target_language": {
                                        "description": "Target language code",
                                        "type": "string",
                                        "example": "hi"
                                    },
                                    "source_language": {
                                        "description": "Source language code (optional)",
                                        "type": "string",
                                        "example": "en"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Translation completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "text": {
                                                    "type": "string",
                                                    "example": "नमस्ते, आप कैसे हैं?"
                                                },
                                                "source_language": {
                                                    "type": "string",
                                                    "example": "en"
                                                },
                                                "target_language": {
                                                    "type": "string",
                                                    "example": "hi"
                                                },
                                                "translated": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "confidence": {
                                                    "type": "number",
                                                    "example": 0.95
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ai/translate-batch": {
            "post": {
                "tags": [
                    "Language Support"
                ],
                "summary": "Translate multiple texts",
                "description": "Translate multiple texts in a single request",
                "operationId": "46ab2f0dc878c2bbc861e2988a060277",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "texts",
                                    "target_language"
                                ],
                                "properties": {
                                    "texts": {
                                        "description": "Texts to translate",
                                        "type": "object",
                                        "example": {
                                            "greeting": "Hello",
                                            "farewell": "Goodbye"
                                        }
                                    },
                                    "target_language": {
                                        "description": "Target language code",
                                        "type": "string",
                                        "example": "hi"
                                    },
                                    "source_language": {
                                        "description": "Source language code (optional)",
                                        "type": "string",
                                        "example": "en"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Batch translation completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "greeting": {
                                                    "properties": {
                                                        "success": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "text": {
                                                            "type": "string",
                                                            "example": "नमस्ते"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "farewell": {
                                                    "properties": {
                                                        "success": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "text": {
                                                            "type": "string",
                                                            "example": "अलविदा"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ai/detect-language": {
            "post": {
                "tags": [
                    "Language Support"
                ],
                "summary": "Detect language of text",
                "description": "Detect the language of input text using Google Translate API",
                "operationId": "69cde1be2d92cc00ca2e8f351bd285d0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "description": "Text to detect language for",
                                        "type": "string",
                                        "example": "नमस्ते, आप कैसे हैं?"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Language detection completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "language": {
                                                    "type": "string",
                                                    "example": "hi"
                                                },
                                                "confidence": {
                                                    "type": "number",
                                                    "example": 0.95
                                                },
                                                "language_name": {
                                                    "type": "string",
                                                    "example": "Hindi"
                                                },
                                                "native_name": {
                                                    "type": "string",
                                                    "example": "हिन्दी"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}